It works!
But just barely! It's very dumb at the moment. And it appears to not work at
all when coupled wit LuaJIT, or at least work in a very mysterious manner not
at all related to the logic I think I wrote.

Very messy. I need to bin the whole thing and start from scratch. But that's
ok. :shipit: